fixed issues with pyinstaller
make sure you turn the console off in the spec file, i used it to debug.

one of the issues was the spec file was still formated incorrectly, another was with the resource_path variable, I added a simplier way for python to find the files in the temp folder
