+++ MyProject/Newnew.spec
@@ -1,0 +1,34 @@
+# -*- mode: python -*-
+
+block_cipher = None
+
+
+added_files = [
+         ( "bcgr.png", '.'),                           
+         ]
+
+
+a = Analysis(['newnew.py'],
+             pathex=['D:\\MyProject'],
+             binaries=[],
+             datas=added_files,
+             hiddenimports=[],
+             hookspath=[],
+             runtime_hooks=[],
+             excludes=[],
+             win_no_prefer_redirects=False,
+             win_private_assemblies=False,
+             cipher=block_cipher)
+pyz = PYZ(a.pure, a.zipped_data,
+             cipher=block_cipher)
+exe = EXE(pyz,
+          a.scripts,
+          a.binaries,
+          a.zipfiles,
+          a.datas,
+          name='newnew',
+          debug=False,
+          strip=False,
+          upx=True,
+          runtime_tmpdir=None,
+          console=True , icon='my_icon.ico')
